# backend/src/main/java/com/meridian/lms/dto/response/UserResponse.java
package com.meridian.lms.dto.response;

import com.meridian.lms.entity.User;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class UserResponse {
    private Long id;
    private String email;
    private String fullName;
    private String phone;
    private String role;
    private Integer creditScore;
    private String status;
    private LocalDateTime createdAt;

    public static UserResponse from(User user) {
        return UserResponse.builder()
                .id(user.getId())
                .email(user.getEmail())
                .fullName(user.getFullName())
                .phone(user.getPhone())
                .role(user.getRole().name())
                .creditScore(user.getCreditScore())
                .status(user.getStatus().name())
                .createdAt(user.getCreatedAt())
                .build();
    }
}