# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"Meridian Loan Management System API","description":"REST API for the Meridian LMS — a loan management platform\nwith customer portal, admin dashboard, credit scoring engine,\nand analytics.\n","contact":{"name":"Meridian Engineering","email":"api@meridian.dev"},"license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"},"version":"v1"},"servers":[{"url":"http://localhost:4000","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Authentication","description":"Register, login, and manage user sessions"}],"paths":{"/api/v1/loans/{id}/review":{"put":{"tags":["loan-controller"],"operationId":"review","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewLoanRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoanResponse"}}}}}}},"/api/v1/employees/{id}":{"get":{"tags":["employee-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}}}},"put":{"tags":["employee-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEmployeeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}}}},"delete":{"tags":["employee-controller"],"operationId":"deactivate","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/employees/{id}/permissions":{"put":{"tags":["employee-controller"],"operationId":"updatePermissions","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePermissionsRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}}}}},"/api/v1/customers/{id}/suspend":{"put":{"tags":["customer-controller"],"operationId":"suspend","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerResponse"}}}}}}},"/api/v1/customers/{id}/activate":{"put":{"tags":["customer-controller"],"operationId":"activate","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerResponse"}}}}}}},"/api/v1/payments/pay":{"post":{"tags":["payment-controller"],"operationId":"pay","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentResponse"}}}}}}},"/api/v1/loans":{"post":{"tags":["loan-controller"],"operationId":"apply","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoanApplicationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoanResponse"}}}}}}},"/api/v1/employees":{"get":{"tags":["employee-controller"],"operationId":"list","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeResponse"}}}}}}},"post":{"tags":["employee-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEmployeeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EmployeeResponse"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["Authentication"],"summary":"Register a new customer","description":"Creates a new customer account and returns a JWT token.","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"Registration successful","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}},"400":{"description":"Invalid input or email already registered","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Authentication"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/api/v1/payments/loan/{loanId}":{"get":{"tags":["payment-controller"],"operationId":"listForLoan","parameters":[{"name":"loanId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaymentResponse"}}}}}}}},"/api/v1/loans/{id}":{"get":{"tags":["loan-controller"],"operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoanResponse"}}}}}}},"/api/v1/loans/types":{"get":{"tags":["loan-type-controller"],"operationId":"list_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LoanTypeResponse"}}}}}}}},"/api/v1/loans/my":{"get":{"tags":["loan-controller"],"operationId":"myLoans","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LoanResponse"}}}}}}}},"/api/v1/health":{"get":{"tags":["health-controller"],"operationId":"health","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{}}}}}}}},"/api/v1/customers":{"get":{"tags":["customer-controller"],"operationId":"list_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CustomerResponse"}}}}}}}},"/api/v1/customers/{id}":{"get":{"tags":["customer-controller"],"operationId":"getById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CustomerResponse"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["Authentication"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/api/v1/analytics/trends":{"get":{"tags":["analytics-controller"],"operationId":"trends","parameters":[{"name":"months","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":12}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TrendPoint"}}}}}}}},"/api/v1/analytics/revenue":{"get":{"tags":["analytics-controller"],"operationId":"revenue","parameters":[{"name":"months","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":12}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RevenuePoint"}}}}}}}},"/api/v1/analytics/loan-types":{"get":{"tags":["analytics-controller"],"operationId":"loanTypes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LoanTypeBreakdown"}}}}}}}},"/api/v1/analytics/dashboard":{"get":{"tags":["analytics-controller"],"operationId":"dashboard","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DashboardKpiResponse"}}}}}}}},"components":{"schemas":{"ReviewLoanRequest":{"type":"object","properties":{"decision":{"type":"string","minLength":1,"pattern":"APPROVE|REJECT|ASSIGN|START_REVIEW"},"reason":{"type":"string","maxLength":1000,"minLength":0}},"required":["decision"]},"LoanResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"customerId":{"type":"integer","format":"int64"},"customerName":{"type":"string"},"customerEmail":{"type":"string"},"loanTypeId":{"type":"integer","format":"int64"},"loanTypeName":{"type":"string"},"loanTypeCode":{"type":"string"},"assignedEmployeeId":{"type":"integer","format":"int64"},"assignedEmployeeName":{"type":"string"},"amount":{"type":"number"},"termMonths":{"type":"integer","format":"int32"},"interestRate":{"type":"number"},"monthlyPayment":{"type":"number"},"totalPayable":{"type":"number"},"totalInterest":{"type":"number"},"outstandingBalance":{"type":"number"},"status":{"type":"string"},"purpose":{"type":"string"},"rejectionReason":{"type":"string"},"appliedAt":{"type":"string","format":"date-time"},"reviewedAt":{"type":"string","format":"date-time"},"approvedAt":{"type":"string","format":"date-time"},"completedAt":{"type":"string","format":"date-time"}}},"UpdateEmployeeRequest":{"type":"object","properties":{"fullName":{"type":"string","maxLength":255,"minLength":2},"phone":{"type":"string","maxLength":50,"minLength":0}}},"EmployeeResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"fullName":{"type":"string"},"phone":{"type":"string"},"status":{"type":"string"},"permissions":{"type":"array","items":{"type":"string"},"uniqueItems":true},"createdAt":{"type":"string","format":"date-time"}}},"UpdatePermissionsRequest":{"type":"object","properties":{"permissions":{"type":"array","items":{"type":"string"},"uniqueItems":true}},"required":["permissions"]},"CustomerResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"fullName":{"type":"string"},"phone":{"type":"string"},"creditScore":{"type":"integer","format":"int32"},"status":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"PaymentRequest":{"type":"object","properties":{"paymentId":{"type":"integer","format":"int64"},"paymentMethod":{"type":"string","maxLength":50,"minLength":0},"idempotencyKey":{"type":"string","maxLength":100,"minLength":8},"amount":{"type":"number"}},"required":["idempotencyKey","paymentId","paymentMethod"]},"PaymentResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"loanId":{"type":"integer","format":"int64"},"amount":{"type":"number"},"principalPortion":{"type":"number"},"interestPortion":{"type":"number"},"dueDate":{"type":"string","format":"date"},"paidAt":{"type":"string","format":"date-time"},"status":{"type":"string"},"paymentMethod":{"type":"string"},"lateFee":{"type":"number"},"installmentNumber":{"type":"integer","format":"int32"}}},"LoanApplicationRequest":{"type":"object","properties":{"loanTypeCode":{"type":"string","minLength":1},"amount":{"type":"number","minimum":0.01},"termMonths":{"type":"integer","format":"int32","maximum":360,"minimum":1},"purpose":{"type":"string","maxLength":1000,"minLength":0}},"required":["amount","loanTypeCode","termMonths"]},"CreateEmployeeRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","maxLength":255,"minLength":0},"password":{"type":"string","maxLength":100,"minLength":8},"fullName":{"type":"string","maxLength":255,"minLength":2},"phone":{"type":"string","maxLength":50,"minLength":0},"permissions":{"type":"array","items":{"type":"string"},"uniqueItems":true}},"required":["email","fullName","password"]},"RegisterRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","maxLength":255,"minLength":0},"password":{"type":"string","maxLength":100,"minLength":8,"pattern":"^(?=.*[A-Za-z])(?=.*\\d).+$"},"fullName":{"type":"string","maxLength":255,"minLength":2},"phone":{"type":"string","pattern":"^[+]?[0-9\\-\\s()]{7,20}$"}},"required":["email","fullName","password"]},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"},"tokenType":{"type":"string"},"expiresInMs":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/UserResponse"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"fullName":{"type":"string"},"phone":{"type":"string"},"role":{"type":"string"},"creditScore":{"type":"integer","format":"int32"},"status":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"LoanTypeResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"code":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"minAmount":{"type":"number"},"maxAmount":{"type":"number"},"minTermMonths":{"type":"integer","format":"int32"},"maxTermMonths":{"type":"integer","format":"int32"},"baseInterestRate":{"type":"number"}}},"TrendPoint":{"type":"object","properties":{"period":{"type":"string"},"applications":{"type":"integer","format":"int64"},"approvals":{"type":"integer","format":"int64"},"rejections":{"type":"integer","format":"int64"}}},"RevenuePoint":{"type":"object","properties":{"period":{"type":"string"},"interestRevenue":{"type":"number"},"lateFees":{"type":"number"},"total":{"type":"number"}}},"LoanTypeBreakdown":{"type":"object","properties":{"code":{"type":"string"},"name":{"type":"string"},"count":{"type":"integer","format":"int64"},"totalDisbursed":{"type":"number"},"approvalRate":{"type":"number","format":"double"}}},"DashboardKpiResponse":{"type":"object","properties":{"totalLoans":{"type":"integer","format":"int64"},"pendingLoans":{"type":"integer","format":"int64"},"activeLoans":{"type":"integer","format":"int64"},"completedLoans":{"type":"integer","format":"int64"},"rejectedLoans":{"type":"integer","format":"int64"},"totalCustomers":{"type":"integer","format":"int64"},"totalDisbursed":{"type":"number"},"totalRevenue":{"type":"number"},"totalOutstanding":{"type":"number"},"approvalRate":{"type":"number","format":"double"}}}},"securitySchemes":{"bearerAuth":{"type":"http","description":"Provide the JWT token from /api/v1/auth/login","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}